+++ test-gatsby/CHANGELOG.md
@@ -1,5 +1,19 @@
 # test-gatsby-thing
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`361a34f`](https://github.com/Thinkmill/manypkg/commit/361a34faac94f7a954bbe00321647fc99ae76c17)]:
+  - @manypkg/gatsby-source-workspace@0.5.1
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [[`f046017`](https://github.com/Thinkmill/manypkg/commit/f046017af2349f0c1bbc5b25224da0ede8ddc2d6)]:
+  - @manypkg/gatsby-source-workspace@0.5.0
+
 ## 0.0.5
 
 ### Patch Changes